+++ CHANGELOG.md
@@ -1,6 +1,16 @@
+## [10.3.2] - 2020-01-09
+
+## Added
+- Editable Table stories
+
+## Changed
+- Export Editable Table component
+
 ## [10.3.1] - 2020-01-09
+
 ## Changed
 - Highlight total columns in ReportTable
+
 ## [10.3.0] - 2020-01-07
 
 ## Added